+++ src/Symfony/Bridge/Twig/Tests/Extension/HttpKernelExtensionTest.php
@@ -11,13 +11,14 @@
 
 namespace Symfony\Bridge\Twig\Tests\Extension;
 
+use PHPUnit\Framework\TestCase;
 use Symfony\Bridge\Twig\Extension\HttpKernelExtension;
 use Symfony\Bridge\Twig\Extension\HttpKernelRuntime;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\HttpKernel\Fragment\FragmentHandler;
 
-class HttpKernelExtensionTest extends \PHPUnit_Framework_TestCase
+class HttpKernelExtensionTest extends TestCase
 {
     /**
      * @expectedException \Twig_Error_Runtime
@@ -63,9 +64,7 @@ class HttpKernelExtensionTest extends \PHPUnit_Framework_TestCase
 
         $context->expects($this->any())->method('getCurrentRequest')->will($this->returnValue(Request::create('/')));
 
-        $renderer = new FragmentHandler($context, array($strategy), false);
-
-        return $renderer;
+        return new FragmentHandler($context, array($strategy), false);
     }
 
     protected function renderTemplate(FragmentHandler $renderer, $template = '{{ render("foo") }}')