+++ src/Symfony/Component/DependencyInjection/Loader/FileLoader.php
@@ -11,14 +11,12 @@
 
 namespace Symfony\Component\DependencyInjection\Loader;
 
+use Symfony\Component\DependencyInjection\ChildDefinition;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\DependencyInjection\Definition;
 use Symfony\Component\DependencyInjection\Exception\InvalidArgumentException;
-use Symfony\Component\DependencyInjection\Exception\LogicException;
 use Symfony\Component\Config\Loader\FileLoader as BaseFileLoader;
 use Symfony\Component\Config\FileLocatorInterface;
-use Symfony\Component\Finder\Finder;
-use Symfony\Component\Finder\Glob;
 
 /**
  * FileLoader is the abstract class used by all built-in loaders that are file based.
@@ -28,6 +26,8 @@ use Symfony\Component\Finder\Glob;
 abstract class FileLoader extends BaseFileLoader
 {
     protected $container;
+    protected $isLoadingInstanceof = false;
+    protected $instanceof = array();
 
     /**
      * @param ContainerBuilder     $container A ContainerBuilder instance
@@ -63,7 +63,22 @@ abstract class FileLoader extends BaseFileLoader
         $prototype = serialize($prototype);
 
         foreach ($classes as $class) {
-            $this->container->setDefinition($class, unserialize($prototype));
+            $this->setDefinition($class, unserialize($prototype));
+        }
+    }
+
+    /**
+     * @experimental in version 3.3
+     */
+    protected function setDefinition($id, Definition $definition)
+    {
+        if ($this->isLoadingInstanceof) {
+            if (!$definition instanceof ChildDefinition) {
+                throw new InvalidArgumentException(sprintf('Invalid type definition "%s": ChildDefinition expected, "%s" given.', $id, get_class($definition)));
+            }
+            $this->instanceof[$id] = $definition;
+        } else {
+            $this->container->setDefinition($id, $definition->setInstanceofConditionals($this->instanceof));
         }
     }
 
@@ -71,9 +86,13 @@ abstract class FileLoader extends BaseFileLoader
     {
         $classes = array();
         $extRegexp = defined('HHVM_VERSION') ? '/\\.(?:php|hh)$/' : '/\\.php$/';
+        $prefixLen = null;
+        foreach ($this->glob($resource, true, $prefix) as $path => $info) {
+            if (null === $prefixLen) {
+                $prefixLen = strlen($prefix);
+            }
 
-        foreach ($this->glob($resource, true, $prefixLen) as $path => $info) {
-            if (!preg_match($extRegexp, $path, $m) || !$info->isFile() || !$info->isReadable()) {
+            if (!preg_match($extRegexp, $path, $m) || !$info->isReadable()) {
                 continue;
             }
             $class = $namespace.ltrim(str_replace('/', '\\', substr($path, $prefixLen, -strlen($m[0]))), '\\');
@@ -89,58 +108,11 @@ abstract class FileLoader extends BaseFileLoader
             }
         }
 
+        if (null !== $prefix) {
+            // track directories only for new & removed files
+            $this->container->fileExists($prefix, '/^$/');
+        }
+
         return $classes;
-    }
-
-    private function glob($resource, $recursive, &$prefixLen = null)
-    {
-        if (strlen($resource) === $i = strcspn($resource, '*?{[')) {
-            $resourcePrefix = $resource;
-            $resource = '';
-        } elseif (0 === $i) {
-            $resourcePrefix = '.';
-            $resource = '/'.$resource;
-        } else {
-            $resourcePrefix = dirname(substr($resource, 0, 1 + $i));
-            $resource = substr($resource, strlen($resourcePrefix));
-        }
-
-        $resourcePrefix = $this->locator->locate($resourcePrefix, $this->currentDir, true);
-        $resourcePrefix = realpath($resourcePrefix) ?: $resourcePrefix;
-        $prefixLen = strlen($resourcePrefix);
-
-        // track directories only for new & removed files
-        $this->container->fileExists($resourcePrefix, '/^$/');
-
-        if (false === strpos($resource, '/**/') && (defined('GLOB_BRACE') || false === strpos($resource, '{'))) {
-            foreach (glob($resourcePrefix.$resource, defined('GLOB_BRACE') ? GLOB_BRACE : 0) as $path) {
-                if ($recursive && is_dir($path)) {
-                    $flags = \FilesystemIterator::SKIP_DOTS | \FilesystemIterator::FOLLOW_SYMLINKS;
-                    foreach (new \RecursiveIteratorIterator(new \RecursiveDirectoryIterator($path, $flags)) as $path => $info) {
-                        yield $path => $info;
-                    }
-                } else {
-                    yield $path => new \SplFileInfo($path);
-                }
-            }
-
-            return;
-        }
-
-        if (!class_exists(Finder::class)) {
-            throw new LogicException(sprintf('Extended glob pattern "%s" cannot be used as the Finder component is not installed.', $resource));
-        }
-
-        $finder = new Finder();
-        $regex = Glob::toRegex($resource);
-        if ($recursive) {
-            $regex = substr_replace($regex, '(/|$)', -2, 1);
-        }
-
-        foreach ($finder->followLinks()->in($resourcePrefix) as $path => $info) {
-            if (preg_match($regex, substr($path, $prefixLen))) {
-                yield $path => $info;
-            }
-        }
     }
 }