+++ src/Symfony/Component/Validator/Tests/Mapping/ClassMetadataTest.php
@@ -11,15 +11,15 @@
 
 namespace Symfony\Component\Validator\Tests\Mapping;
 
+use PHPUnit\Framework\TestCase;
 use Symfony\Component\Validator\Constraint;
-use Symfony\Component\Validator\Constraints\GreaterThan;
 use Symfony\Component\Validator\Constraints\Valid;
 use Symfony\Component\Validator\Mapping\ClassMetadata;
 use Symfony\Component\Validator\Tests\Fixtures\ConstraintA;
 use Symfony\Component\Validator\Tests\Fixtures\ConstraintB;
 use Symfony\Component\Validator\Tests\Fixtures\PropertyConstraint;
 
-class ClassMetadataTest extends \PHPUnit_Framework_TestCase
+class ClassMetadataTest extends TestCase
 {
     const CLASSNAME = 'Symfony\Component\Validator\Tests\Fixtures\Entity';
     const PARENTCLASS = 'Symfony\Component\Validator\Tests\Fixtures\EntityParent';
@@ -303,21 +303,6 @@ class ClassMetadataTest extends \PHPUnit_Framework_TestCase
     public function testGetPropertyMetadataReturnsEmptyArrayWithoutConfiguredMetadata()
     {
         $this->assertCount(0, $this->metadata->getPropertyMetadata('foo'), '->getPropertyMetadata() returns an empty collection if no metadata is configured for the given property');
-    }
-
-    public function testMergeDoesOverrideConstraintsFromParentClassIfPropertyIsOverriddenInChildClass()
-    {
-        $parentMetadata = new ClassMetadata('\Symfony\Component\Validator\Tests\Mapping\ParentClass');
-        $parentMetadata->addPropertyConstraint('example', new GreaterThan(0));
-
-        $childMetadata = new ClassMetadata('\Symfony\Component\Validator\Tests\Mapping\ChildClass');
-        $childMetadata->addPropertyConstraint('example', new GreaterThan(1));
-        $childMetadata->mergeConstraints($parentMetadata);
-
-        $expectedMetadata = new ClassMetadata('\Symfony\Component\Validator\Tests\Mapping\ChildClass');
-        $expectedMetadata->addPropertyConstraint('example', new GreaterThan(1));
-
-        $this->assertEquals($expectedMetadata, $childMetadata);
     }
 }